Add our first integration test, currently failing
This spec uses Open3 to execute the ruby script from the outside. The
first failing assertion is ensuring that an argument is provided
defining the exit status of the script.

Open3 is used as it allows you to capture the process exit status, as
well as the stdout and stderr handles.