Add failing test for basic functionality
These failing tests describe the basic functionality we want the
application to have.

At this stage, I'm ignoring the requirement for uniques (I have a
Plan(tm) for how to build that in later)

This should let me start writing the framework of my app (probably in
/lib) as classes which I can unit test in isolation.

A single additional error handling case has been implemented here; I
want the script to verify that the file exists before handing it down to
the implementing ruby classes; that way I don't have to handle file
access errors further down the stack.