Add failing tests to #each_by_hits
This spec covers many assertions; the repository is the meat of the
"implementation" of this technical assessment.

It's going to handle the sorting, indexing, and so on -- but the good
news is if we implement it naively now we'll be able to refactor later.

To make this test pass, we'll need to implement some form of data
structure to be then subsequently sorted. Quite a bit of code coming in
the next commit I think.